remove dead code and reorganize
