+++ fiftyone/server/routes/sample.py
@@ -11,7 +11,6 @@ gate on, and the new value to write. A failed precondition returns ``409``
 with the document's current state so the client can reconcile.
 """
 
-import datetime
 import logging
 from collections import defaultdict
 from dataclasses import dataclass
@@ -27,21 +26,20 @@ from starlette.requests import Request
 import fiftyone.core.labels as fol
 import fiftyone.core.utils as fou
 from fiftyone.core.odm.database import get_db_conn
-from fiftyone.server import decorators, utils
+from fiftyone.server import decorators
+from fiftyone.server.annotations import (
+    build_label_update,
+    build_write,
+    is_label,
+)
 from fiftyone.server.utils.json.encoder import JSONResponse
 
 logger = logging.getLogger(__name__)
 
 
 # ---------------------------------------------------------------------------
-# Value coercion and gating
+# Planning and application
 # ---------------------------------------------------------------------------
-
-
-def _now() -> datetime.datetime:
-    # Naive UTC to match how fiftyone stamps ``last_modified_at``; a tz-aware
-    # value breaks naive/aware comparisons in code that reads the field.
-    return datetime.datetime.utcnow()
 
 
 def _object_id(value: Any) -> ObjectId:
@@ -57,281 +55,13 @@ def _object_id(value: Any) -> ObjectId:
         ) from err
 
 
-def _is_label(value: Any) -> bool:
-    return isinstance(value, dict) and "_cls" in value
-
-
-def _label_to_mongo(value: dict) -> dict:
-    """Coerces a label dict to its stored BSON form (pure CPU, no DB access).
-
-    Round-trips through the label class so masks, datetimes, ObjectIds, etc.
-    are encoded exactly as fiftyone persists them — required for equality
-    gates to match.
-    """
-    try:
-        return utils.json.deserialize(value).to_mongo()
-    except Exception as err:  # pylint: disable=broad-except
-        raise HTTPException(
-            status_code=400, detail=f"Could not deserialize label: {err}"
-        ) from err
-
-
-def _gatable(value: Any) -> bool:
-    """Whether ``value`` is safe to use as an equality precondition.
-
-    Embedded documents are field-order-sensitive and binary blobs may not
-    re-serialize byte-for-byte, so those gate on identity only.
-    """
+def _required(update: dict, key: str) -> Any:
+    value = update.get(key)
     if value is None:
-        return True
-    if isinstance(value, (dict, bytes, bytearray)):
-        return False
-    if isinstance(value, (list, tuple)):
-        return all(_gatable(v) for v in value)
-    return True
-
-
-def _gate(value: Any) -> Any:
-    # ``$in: [None]`` matches both an absent field and an explicit null.
-    return {"$in": [None]} if value is None else value
-
-
-def _scalar_gates(bson_doc: dict) -> Dict[str, Any]:
-    return {
-        field: _gate(value)
-        for field, value in bson_doc.items()
-        if field != "_id" and _gatable(value)
-    }
-
-
-def _changed_label_fields(
-    old_bson: dict, new_bson: dict
-) -> Tuple[Dict[str, Any], List[str]]:
-    """Returns ``(set_fields, unset_fields)`` for a label modification;
-    ``_id`` is never touched."""
-    set_fields: Dict[str, Any] = {}
-    unset_fields: List[str] = []
-    for field in set(old_bson) | set(new_bson):
-        if field == "_id":
-            continue
-        old_val = old_bson.get(field)
-        new_val = new_bson.get(field)
-        if old_val == new_val:
-            continue
-        if field in new_bson:
-            set_fields[field] = new_val
-        else:
-            unset_fields.append(field)
-    return set_fields, unset_fields
-
-
-# ---------------------------------------------------------------------------
-# Gated write construction
-# ---------------------------------------------------------------------------
-
-
-def _build_label_update(
-    lookup_path: str,
-    label_oid: Optional[ObjectId],
-    old: Optional[dict],
-    new: Optional[dict],
-    *,
-    array: bool,
-    doc_filter: Dict[str, Any],
-) -> Tuple[Dict[str, Any], Dict[str, Any]]:
-    """Builds the (filter, update) for an add/remove/modify of one label.
-
-    ``array`` selects between a label inside a list field (``$elemMatch`` +
-    positional ``$``) and a flattened single label (matched by path).
-    """
-    now = _now()
-
-    if new is None:
-        # Gate on the scalars the editor saw so a stale delete can't erase a
-        # label another editor just changed.
-        old_bson = _label_to_mongo(old) if old is not None else {}
-        gates = _scalar_gates(old_bson)
-        if array:
-            filter_doc = {
-                **doc_filter,
-                lookup_path: {"$elemMatch": {"_id": label_oid, **gates}},
-            }
-            update_doc = {
-                "$pull": {lookup_path: {"_id": label_oid}},
-                "$set": {"last_modified_at": now},
-            }
-        else:
-            filter_doc = {**doc_filter}
-            # Bind to identity so a replacement label with equal scalars but a
-            # new _id isn't mistaken for ours.
-            if "_id" in old_bson:
-                filter_doc[f"{lookup_path}._id"] = old_bson["_id"]
-            filter_doc.update(
-                {f"{lookup_path}.{f}": v for f, v in gates.items()}
-            )
-            update_doc = {
-                "$unset": {lookup_path: ""},
-                "$set": {"last_modified_at": now},
-            }
-        return filter_doc, update_doc
-
-    new_bson = _label_to_mongo(new)
-
-    if old is None:
-        if array:
-            filter_doc = {
-                **doc_filter,
-                f"{lookup_path}._id": {"$ne": label_oid},
-            }
-            update_doc = {
-                "$push": {lookup_path: new_bson},
-                "$set": {"last_modified_at": now},
-            }
-        else:
-            # Create only while still absent/null so a stale create can't
-            # clobber a value another editor inserted.
-            filter_doc = {**doc_filter, lookup_path: {"$in": [None]}}
-            update_doc = {
-                "$set": {lookup_path: new_bson, "last_modified_at": now}
-            }
-        return filter_doc, update_doc
-
-    old_bson = _label_to_mongo(old)
-    set_fields, unset_fields = _changed_label_fields(old_bson, new_bson)
-
-    prefix = f"{lookup_path}.$." if array else f"{lookup_path}."
-    gates = {
-        field: _gate(old_bson.get(field))
-        for field in list(set_fields) + unset_fields
-        if _gatable(old_bson.get(field))
-    }
-
-    if array:
-        filter_doc = {
-            **doc_filter,
-            lookup_path: {"$elemMatch": {"_id": label_oid, **gates}},
-        }
-    else:
-        filter_doc = {**doc_filter}
-        if "_id" in old_bson:
-            filter_doc[f"{lookup_path}._id"] = old_bson["_id"]
-        filter_doc.update({f"{lookup_path}.{f}": v for f, v in gates.items()})
-
-    set_doc = {f"{prefix}{f}": v for f, v in set_fields.items()}
-    set_doc["last_modified_at"] = now
-    update_doc: Dict[str, Any] = {"$set": set_doc}
-    if unset_fields:
-        update_doc["$unset"] = {f"{prefix}{f}": "" for f in unset_fields}
-
-    return filter_doc, update_doc
-
-
-def _build_label_applied_filter(
-    lookup_path: str,
-    label_oid: Optional[ObjectId],
-    new: Optional[dict],
-    *,
-    array: bool,
-    doc_filter: Dict[str, Any],
-) -> Dict[str, Any]:
-    """A filter matching the document iff the label update already holds —
-    judged on identity plus gatable (scalar) fields of the new value."""
-    if new is None:
-        if array:
-            return {**doc_filter, f"{lookup_path}._id": {"$ne": label_oid}}
-        return {**doc_filter, lookup_path: {"$in": [None]}}
-
-    new_bson = _label_to_mongo(new)
-    gates = _scalar_gates(new_bson)
-
-    if array:
-        return {
-            **doc_filter,
-            lookup_path: {"$elemMatch": {"_id": label_oid, **gates}},
-        }
-
-    filter_doc = {**doc_filter}
-    if "_id" in new_bson:
-        filter_doc[f"{lookup_path}._id"] = new_bson["_id"]
-    filter_doc.update({f"{lookup_path}.{f}": v for f, v in gates.items()})
-    return filter_doc
-
-
-def _build_primitive_update(
-    field_path: str,
-    old: Any,
-    new: Any,
-    *,
-    doc_filter: Dict[str, Any],
-) -> Tuple[Dict[str, Any], Dict[str, Any]]:
-    now = _now()
-    filter_doc = dict(doc_filter)
-    if _gatable(old):
-        filter_doc[field_path] = _gate(old)
-
-    if new is None:
-        update_doc = {
-            "$unset": {field_path: ""},
-            "$set": {"last_modified_at": now},
-        }
-    else:
-        update_doc = {"$set": {field_path: new, "last_modified_at": now}}
-
-    return filter_doc, update_doc
-
-
-def _build_primitive_applied_filter(
-    field_path: str,
-    new: Any,
-    *,
-    doc_filter: Dict[str, Any],
-) -> Optional[Dict[str, Any]]:
-    """A filter matching iff the primitive update already holds, or ``None``
-    when the new value is not reliably comparable (non-scalar)."""
-    if new is None:
-        return {**doc_filter, field_path: {"$in": [None]}}
-    if _gatable(new):
-        return {**doc_filter, field_path: new}
-    return None
-
-
-def _build_write(
-    lookup_path: str,
-    label_oid: Optional[ObjectId],
-    old: Any,
-    new: Any,
-    doc_filter: Dict[str, Any],
-) -> Tuple[Dict[str, Any], Dict[str, Any], Optional[Dict[str, Any]]]:
-    """(filter, update, applied_filter) for one gated field write — a label
-    list element, a flat label, or a primitive."""
-    if label_oid is not None:
-        filter_doc, update_doc = _build_label_update(
-            lookup_path, label_oid, old, new, array=True, doc_filter=doc_filter
+        raise HTTPException(
+            status_code=400, detail=f"Update is missing '{key}'"
         )
-        applied_filter = _build_label_applied_filter(
-            lookup_path, label_oid, new, array=True, doc_filter=doc_filter
-        )
-    elif _is_label(new) or _is_label(old):
-        filter_doc, update_doc = _build_label_update(
-            lookup_path, None, old, new, array=False, doc_filter=doc_filter
-        )
-        applied_filter = _build_label_applied_filter(
-            lookup_path, None, new, array=False, doc_filter=doc_filter
-        )
-    else:
-        filter_doc, update_doc = _build_primitive_update(
-            lookup_path, old, new, doc_filter=doc_filter
-        )
-        applied_filter = _build_primitive_applied_filter(
-            lookup_path, new, doc_filter=doc_filter
-        )
-
-    return filter_doc, update_doc, applied_filter
-
-
-# ---------------------------------------------------------------------------
-# Planning and application
-# ---------------------------------------------------------------------------
+    return value
 
 
 @dataclass
@@ -390,15 +120,6 @@ def _is_generated_collection(name: str) -> bool:
     return name.startswith(_GENERATED_COLLECTION_PREFIXES)
 
 
-def _required(update: dict, key: str) -> Any:
-    value = update.get(key)
-    if value is None:
-        raise HTTPException(
-            status_code=400, detail=f"Update is missing '{key}'"
-        )
-    return value
-
-
 def _allowed_collections(db, path_params: dict) -> set:
     """The route dataset's collections (samples + frames), resolved from its
     dataset document. Resolving (rather than constructing names from the id)
@@ -431,7 +152,7 @@ def _extract_mask_write(new: Any) -> Tuple[Any, Optional[Tuple[str, Any]]]:
     stripped from the document write.
     """
     if not (
-        _is_label(new)
+        is_label(new)
         and new.get("mask") is not None
         and isinstance(new.get("mask_path"), str)
         and new["mask_path"]
@@ -491,7 +212,7 @@ def _plan_update(db, update: dict, allowed: set) -> _UpdatePlan:
     label_id = update.get("labelId")
     new, mask_write = _extract_mask_write(update["newValue"])
     doc_filter = {"_id": _object_id(doc_id)}
-    filter_doc, update_doc, applied_filter = _build_write(
+    filter_doc, update_doc, applied_filter = build_write(
         lookup_path,
         _object_id(label_id) if label_id is not None else None,
         update["previousValue"],
@@ -569,7 +290,7 @@ def _plan_generated_sync(
 
     # Flat label / primitive field: the generated copy has the same shape.
     if label_id is None:
-        plan.filter_doc, plan.update_doc, _ = _build_write(
+        plan.filter_doc, plan.update_doc, _ = build_write(
             lookup_path, None, old, new, doc_filter
         )
         return [plan]
@@ -589,7 +310,7 @@ def _plan_generated_sync(
     if embedded_type in _LIST_LABEL_TYPES:
         # Label list (evaluation patches): mirror the source update.
         plan.lookup_path = f"{gen_field}.{lookup_path.rsplit('.', 1)[-1]}"
-        plan.filter_doc, plan.update_doc = _build_label_update(
+        plan.filter_doc, plan.update_doc = build_label_update(
             plan.lookup_path,
             label_oid,
             old,
@@ -607,7 +328,7 @@ def _plan_generated_sync(
         return [plan]
 
     plan.lookup_path = gen_field
-    plan.filter_doc, plan.update_doc = _build_label_update(
+    plan.filter_doc, plan.update_doc = build_label_update(
         gen_field, None, old, new, array=False, doc_filter=doc_filter
     )
     return [plan]
@@ -691,87 +412,104 @@ class SampleFields(HTTPEndpoint):
             rebase those deltas and retry. Updates not listed in
             ``conflicts`` were applied.
         """
-        updates = data.get("updates") if isinstance(data, dict) else data
-        if not isinstance(updates, list):
-            raise HTTPException(
-                status_code=400, detail="Expected a list of updates"
+        try:
+            updates = data.get("updates") if isinstance(data, dict) else data
+            if not isinstance(updates, list):
+                raise HTTPException(
+                    status_code=400, detail="Expected a list of updates"
+                )
+
+            db = get_db_conn()
+            allowed = _allowed_collections(db, request.path_params)
+
+            logger.debug(
+                "Applying %d field update(s) to sample %s",
+                len(updates),
+                request.path_params.get("sample_id"),
             )
 
-        db = get_db_conn()
-        allowed = _allowed_collections(db, request.path_params)
+            source: List[Tuple[int, _UpdatePlan]] = []
+            generated: List[Tuple[int, _UpdatePlan]] = []
+            for index, update in enumerate(updates):
+                plan = _plan_update(db, update, allowed)
+                source.append((index, plan))
+                for sync in _plan_generated_sync(db, update, plan.lookup_path):
+                    generated.append((index, sync))
 
-        logger.debug(
-            "Applying %d field update(s) to sample %s",
-            len(updates),
-            request.path_params.get("sample_id"),
-        )
+            # Masks destined for disk are written before any database write —
+            # a failure aborts the request while nothing has been applied.
+            for _index, plan in source:
+                if plan.mask_write is not None:
+                    _write_mask_file(*plan.mask_write)
 
-        source: List[Tuple[int, _UpdatePlan]] = []
-        generated: List[Tuple[int, _UpdatePlan]] = []
-        for index, update in enumerate(updates):
-            plan = _plan_update(db, update, allowed)
-            source.append((index, plan))
-            for sync in _plan_generated_sync(db, update, plan.lookup_path):
-                generated.append((index, sync))
+            conflicts = []
+            for index, plan in _apply_bulk(source):
+                # Read-only diagnosis: the update either applied in the bulk
+                # pass / already held (lost-response retry), or missed its
+                # gate. An update changed *again* by a concurrent writer
+                # before this read lands as a conflict too and reconciles
+                # identically. A precondition miss is expected under
+                # concurrent editing, not a fault — the 409 summary below is
+                # the one INFO line; per-update detail stays at debug.
+                if plan.already_holds():
+                    logger.debug(
+                        "Annotation update applied/held (confirmed by "
+                        "read): collection=%s id=%s lookupPath=%s",
+                        plan.collection_name,
+                        plan.doc_id,
+                        plan.lookup_path,
+                    )
+                    continue
 
-        # Masks destined for disk are written before any database write — a
-        # failure aborts the request while nothing has been applied.
-        for _index, plan in source:
-            if plan.mask_write is not None:
-                _write_mask_file(*plan.mask_write)
-
-        conflicts = []
-        for index, plan in _apply_bulk(source):
-            # Read-only diagnosis: the update either applied in the bulk
-            # pass / already held (lost-response retry), or missed its gate.
-            # An update changed *again* by a concurrent writer before this
-            # read lands as a conflict too and reconciles identically.
-            if plan.already_holds():
-                logger.info(
-                    "Annotation update applied/held (confirmed by read): "
+                logger.debug(
+                    "Annotation update rejected (precondition mismatch): "
                     "collection=%s id=%s lookupPath=%s",
                     plan.collection_name,
                     plan.doc_id,
                     plan.lookup_path,
                 )
-                continue
+                conflicts.append(
+                    {"index": index, "value": plan.current_document()}
+                )
 
-            logger.warning(
-                "Annotation save REJECTED (precondition mismatch): "
-                "collection=%s id=%s lookupPath=%s",
-                plan.collection_name,
-                plan.doc_id,
-                plan.lookup_path,
-            )
-            conflicts.append(
-                {"index": index, "value": plan.current_document()}
-            )
+            # Runs even when some source writes conflicted — the applied
+            # ones still want their generated copies in sync.
+            unconfirmed = _apply_bulk(generated, best_effort=True)
+            if unconfirmed:
+                logger.debug(
+                    "Generated-view sync unconfirmed for %d update(s) "
+                    "(precondition mismatch)",
+                    len(unconfirmed),
+                )
 
-        # Runs even when some source writes conflicted — the applied ones
-        # still want their generated copies in sync.
-        unconfirmed = _apply_bulk(generated, best_effort=True)
-        if unconfirmed:
-            logger.info(
-                "Generated-view sync unconfirmed for %d update(s) "
-                "(precondition mismatch)",
-                len(unconfirmed),
-            )
+            if conflicts:
+                logger.info(
+                    "Annotation save for sample %s: %d of %d update(s) "
+                    "conflicted (409)",
+                    request.path_params.get("sample_id"),
+                    len(conflicts),
+                    len(source),
+                )
+                return JSONResponse({"conflicts": conflicts}, status_code=409)
 
-        if conflicts:
-            logger.info(
-                "Annotation save for sample %s: %d of %d update(s) conflicted",
+            logger.debug(
+                "Annotation save for sample %s: %d update(s) applied",
                 request.path_params.get("sample_id"),
-                len(conflicts),
                 len(source),
             )
-            return JSONResponse({"conflicts": conflicts}, status_code=409)
+            return JSONResponse({"updated": len(source)})
 
-        logger.info(
-            "Annotation save for sample %s: %d update(s) applied",
-            request.path_params.get("sample_id"),
-            len(source),
-        )
-        return JSONResponse({"updated": len(source)})
+        except HTTPException as err:
+            # Every rejected (4xx) save gets one server-side log; 5xx are
+            # unexpected and already logged with a traceback where they arise.
+            if err.status_code < 500:
+                logger.info(
+                    "Annotation save for sample %s rejected (%d): %s",
+                    request.path_params.get("sample_id"),
+                    err.status_code,
+                    err.detail,
+                )
+            raise
 
 
 SampleRoutes = [
